integrate razorpay

+++ server/src/controllers/productController.js
@@ -1,9 +1,13 @@
 const productServices = require("../services/productServices.js");
-
+const isValidDocumentId = require("../utils/Validator.js");
+const ExpressError = require("../utils/ExpressError.js");
 
 async function getProductInfo(req, res) {
-
   const { productId } = req.params;
+
+  if (!isValidDocumentId(productId)) {
+    throw new ExpressError(400, "Invalid Product Id");
+  }
 
   const result = await productServices.getProductInfo(productId);
 
@@ -11,15 +15,18 @@ async function getProductInfo(req, res) {
 }
 
 async function getBestSeller(req, res) {
-
   const result = await productServices.getBestSeller();
-  
+
   return res.status(200).json(result);
 }
 
 async function getRelatedProducts(req, res) {
   let { productId } = req.params;
-  
+
+  if (!isValidDocumentId(productId)) {
+    throw new ExpressError(400, "Invalid Product Id");
+  }
+
   const result = await productServices.getRelatedProducts(productId);
 
   return res.status(200).json(result);
@@ -27,18 +34,18 @@ async function getRelatedProducts(req, res) {
 
 async function getAllProducts(req, res) {
   const result = await productServices.getAllProducts();
-  
+
   res.status(200).json(result);
 }
 
 async function latestCollections(req, res) {
   const result = await productServices.latestCollections();
- 
+
   return res.status(200).json(result);
 }
 
 async function addProduct(req, res) {
-  const result = await productServices.addProduct(req.body,req.file);
+  const result = await productServices.addProduct(req.body, req.file);
 
   return res.status(201).json(result);
 }
@@ -46,23 +53,40 @@ async function addProduct(req, res) {
 async function editProductInfo(req, res) {
   const { productId } = req.params;
 
-  const result = await productServices.editProductInfo(req.body,req.file,productId);
+  if (!isValidDocumentId(productId)) {
+    throw new ExpressError(400, "Invalid Product Id");
+  }
+
+  const result = await productServices.editProductInfo(
+    req.body,
+    req.file,
+    productId,
+  );
 
   return res.status(200).json(result);
 }
 
 async function deleteProduct(req, res) {
   const { productId } = req.params;
+
+  if (!isValidDocumentId(productId)) {
+    throw new ExpressError(400, "Invalid Product Id");
+  }
+
   const result = await productServices.deleteProduct(productId);
 
   return res.status(200).json(result);
 }
 
-async function updateProductRating(req,res){
+async function updateProductRating(req, res) {
   const { productId } = req.params;
   const { rating } = req.body;
 
-  const result = await productServices.updateProductRating(productId,rating);
+  if (!isValidDocumentId(productId)) {
+    throw new ExpressError(400, "Invalid Product Id");
+  }
+
+  const result = await productServices.updateProductRating(productId, rating);
 
   return res.status(201).json(result);
 }
@@ -76,5 +100,5 @@ module.exports = {
   getBestSeller,
   getRelatedProducts,
   getAllProducts,
-  updateProductRating
+  updateProductRating,
 };
