+++ server/src/controllers/userController.js
@@ -1,8 +1,10 @@
 const userServices = require("../services/userServices.js");
+const isValidDocumentId = require("../utils/Validator.js");
+const ExpressError = require("../utils/ExpressError.js");
 
 async function getUser(req, res) {
   const result = await userServices.getUser(req.user);
-  
+
   return res.status(200).json(result);
 }
 
@@ -14,24 +16,30 @@ async function getAllUser(req, res) {
 
 async function updateProfileInfo(req, res) {
   const userId = req.user._id;
-  const result = await userServices.updateProfileInfo(req.body,userId);
+  const result = await userServices.updateProfileInfo(req.body, userId);
 
   return res.status(200).json(result);
 }
 
 async function updateProfilePassword(req, res) {
-  const result = await userServices.updateProfilePassword(req.body,req.user);
+  const result = await userServices.updateProfilePassword(req.body, req.user);
 
   return res.status(200).json(result);
 }
 
 async function uploadProfileImage(req, res) {
-  const result = await userServices.uploadProfileImage(req.user,req.file);
+  const result = await userServices.uploadProfileImage(req.user, req.file);
 
   return res.status(200).json(result);
 }
 
 async function deleteUser(req, res) {
+  const { userId } = req.params;
+
+  if (!isValidDocumentId(userId)) {
+    throw new ExpressError(400, "Invalid user Id");
+  }
+
   const result = await userServices.deleteUser(req.params.userId);
 
   return res.status(200).json(result);
